requirements.txt
<<<<<<< HEAD
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
pydantic==1.10.11
python-dotenv==1.0.0
pandas>=2.2.2
openai>=1.0.0
=======
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
pydantic==1.10.11
python-dotenv==1.0.0
pandas==2.0.3
openai>=1.0.0
>>>>>>> 46db45df09567d4ab46114196a8092fea0280438
